DRY refactor card/box checkbox to itemcheckbox
