+++ src/components/Box.vue
@@ -13,6 +13,7 @@ import utils from '@/utils.js'
 import consts from '@/consts.js'
 import fonts from '@/data/fonts.js'
 import ItemConnectorButton from '@/components/ItemConnectorButton.vue'
+import ItemCheckboxButton from '@/components/ItemCheckboxButton.vue'
 import smartquotes from 'smartquotes'
 import postMessage from '@/postMessage.js'
 
@@ -635,30 +636,8 @@ const updateRemoteConnections = () => {
 
 const isChecked = computed(() => utils.nameIsChecked(name.value))
 const hasCheckbox = computed(() => {
-  return utils.checkboxFromString(name.value)
+  return Boolean(utils.checkboxFromString(name.value))
 })
-const checkboxState = computed({
-  get () {
-    return isChecked.value
-  }
-})
-const toggleBoxChecked = () => {
-  if (globalStore.preventDraggedBoxFromShowingDetails) { return }
-  if (!canEditBox.value) { return }
-  const value = !isChecked.value
-  globalStore.closeAllDialogs()
-  boxStore.toggleBoxChecked(props.box.id, value)
-  postMessage.sendHaptics({ name: 'heavyImpact' })
-  cancelLocking()
-  globalStore.currentUserIsDraggingBox = false
-  const userId = userStore.id
-  broadcastStore.update({ updates: { userId }, action: 'clearRemoteBoxesDragging' })
-  event.stopPropagation()
-  globalStore.preventMultipleSelectedActionsIsVisible = false
-  globalStore.clearMultipleSelected()
-  globalStore.currentDraggingBoxId = ''
-  globalStore.updateMultipleBoxesSelectedIds([])
-}
 const containingBoxes = computed(() => {
   if (!state.isVisibleInViewport) { return }
   if (globalStore.currentUserIsDraggingBox) { return }
@@ -741,9 +720,8 @@ const focusColor = computed(() => {
     )
       .locking-frame(v-if="state.isLocking" :style="lockingFrameStyle")
       //- [·]
-      .checkbox-wrap(v-if="hasCheckbox" @mouseup.left="toggleBoxChecked" @touchend.prevent="toggleBoxChecked")
-        label(:class="{active: isChecked, disabled: !canEditBox}")
-          input(name="checkbox" type="checkbox" v-model="checkboxState")
+      ItemCheckboxButton(:visible="hasCheckbox" :box="box" :canEditItem="canEditBox" @toggleItemChecked="cancelLocking")
+      //- name
       .name-wrap(:class="{'is-checked': isChecked}")
         //- simplified nameSegments
         template(v-for="segment in nameSegments")
@@ -891,44 +869,6 @@ const focusColor = computed(() => {
     box-shadow var(--active-shadow)
   &.is-dark
     color var(--primary-on-dark-background)
-
-  .checkbox-wrap
-    padding-left 8px
-    padding-top 5px
-    padding-bottom 6px
-    display inline-block
-    label
-      pointer-events none
-      width 20px
-      height 16px
-      display flex
-      align-items center
-      padding-left 4px
-      padding-right 4px
-      input
-        margin 0
-        margin-top -1px
-        width 10px
-        height 10px
-        background-size contain
-    &:hover
-      label
-        box-shadow 3px 3px 0 var(--heavy-shadow)
-        background-color var(--secondary-hover-background)
-        input
-          background-color var(--secondary-hover-background)
-      label.active
-        box-shadow var(--active-inset-shadow)
-        background-color var(--secondary-active-background)
-        input
-          background-color var(--secondary-active-background)
-    &:active
-      label
-        box-shadow none
-        color var(--primary)
-        background-color var(--secondary-active-background)
-      input
-        background-color var(--secondary-active-background)
 
   .name-wrap
     padding 6px 8px
